azure-underwriting-system/requirements.txt
langgraph>=0.2.40
langchain>=0.2.16
langchain-openai>=0.2.5
langchain-community>=0.2.16
langchain-text-splitters>=0.2.4
chromadb>=0.5.5
pypdf>=4.3.1
python-dotenv>=1.0.1
pydantic>=2.8.2
rich>=13.7.1
fastapi>=0.110.0
